feat(cross data): add link to generic items
- added link to data provider screen to jobs and ads
  - added data provider to generic item query
  - extracted the button and function to navigate to dataprovider screen

SVA-192
