+++ src/components/screens/Tour.js
@@ -1,11 +1,11 @@
 import PropTypes from 'prop-types';
-import React, { useCallback } from 'react';
+import React from 'react';
 import { View } from 'react-native';
 
 import { consts, device, texts } from '../../config';
 import { matomoTrackingString } from '../../helpers';
 import { useMatomoTrackScreenView } from '../../hooks';
-import { Button } from '../Button';
+import { DataProviderButton } from '../DataProviderButton';
 import { HtmlView } from '../HtmlView';
 import { ImageSection } from '../ImageSection';
 import { InfoCard } from '../infoCard';
@@ -62,20 +62,6 @@ export const Tour = ({ data, navigation }) => {
 
   const businessAccount = dataProvider?.dataType === 'business_account';
 
-  const navigateToDataProvider = useCallback(
-    () =>
-      dataProvider &&
-      businessAccount &&
-      navigation.navigate('DataProvider', {
-        dataProviderId: dataProvider.id,
-        dataProviderName: dataProvider.name,
-        logo,
-        title: dataProvider.name
-      }),
-
-    [businessAccount, dataProvider, logo, navigation]
-  );
-
   return (
     <View>
       <ImageSection mediaContents={mediaContents} />
@@ -121,12 +107,7 @@ export const Tour = ({ data, navigation }) => {
         <TMBNotice dataProvider={dataProvider} openWebScreen={openWebScreen} />
 
         {!!businessAccount && (
-          <Wrapper>
-            <Button
-              title={`${texts.dataProvider.more} ${dataProvider.name}`}
-              onPress={navigateToDataProvider}
-            />
-          </Wrapper>
+          <DataProviderButton dataProvider={dataProvider} navigation={navigation} />
         )}
       </WrapperWithOrientation>
     </View>
